continually scan crash directory in background
whoopsie reports successfully uploading a package by creating a
.uploaded file next to the crash report. To be able to reflect that
change in the UI we need to look for that file, so switch everything
around to a model where the crash directory is continually scanned and
changes that are detected are reported to the rest of the code  via a
Queue and signals. It's all a bit out of control but it works.