do not switch snap channel from .disk/info if not on ubuntu/stable-XX.YY
The point of switching snap channel based on .disk/info is so that we
can avoid releasing updates to point release media if they won't work
there. If the snap is not tracking ubuntu/stable-XX.YY then there is
funny business afoot and we are not on point release media, so switching
channels as if we are is just unhelpful (for example, if you are working
on a project that is going to involve giving someone an image with a
subiquity from a non default track).

If the channel to switch too comes from the kernel command line or
autoinstall though, it should still be honoured.