only mount seed from to-be-installed layer if needed
We are going to move to a model where snapd is responsible for copying
the seed to the target system, so we will include it in the live layer
rather than the to-be-copied layer. In this case, we will not need to
mount the seed into the live layer. However we need to retain support
for doing that for backwards compat (for now).